% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_crabs}
\alias{anova_crabs}
\title{My function creates an ANOVA table with the assumption that body depth is predicted by sex of the crabs.}
\usage{
anova_crabs(file.path)
}
\arguments{
\item{file.path}{is the directory to the data set}
}
\value{
Returns the ANOVA table with the assumption that body depth is predicted by sex of the crabs.
}
\description{
My function creates an ANOVA table with the assumption that body depth is predicted by sex of the crabs.
}
